Move committee to it's own package
This improves modularity and clarity.
