+++ src/main/java/ch/wisv/areafiftylan/web/controller/WebRestController.java
@@ -17,20 +17,15 @@
 
 package ch.wisv.areafiftylan.web.controller;
 
-import ch.wisv.areafiftylan.utils.ResponseEntityBuilder;
-import ch.wisv.areafiftylan.web.model.CommitteeMember;
 import ch.wisv.areafiftylan.web.model.Event;
 import ch.wisv.areafiftylan.web.model.Sponsor;
 import ch.wisv.areafiftylan.web.model.Tournament;
 import ch.wisv.areafiftylan.web.service.EventServiceImpl;
 import ch.wisv.areafiftylan.web.service.SponsorServiceImpl;
 import ch.wisv.areafiftylan.web.service.TournamentServiceImpl;
-import ch.wisv.areafiftylan.web.service.WebService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Collection;
 
@@ -42,15 +37,12 @@ public class WebRestController {
     private final EventServiceImpl eventService;
     private final SponsorServiceImpl sponsorService;
 
-    private WebService webService;
-
     @Autowired
     public WebRestController(TournamentServiceImpl tournamentService, EventServiceImpl eventService,
-                             SponsorServiceImpl sponsorService, WebService webService) {
+                             SponsorServiceImpl sponsorService) {
         this.tournamentService = tournamentService;
         this.eventService = eventService;
         this.sponsorService = sponsorService;
-        this.webService = webService;
     }
 
     @RequestMapping("/tournaments")
@@ -69,55 +61,4 @@ public class WebRestController {
     }
 
 
-    /**
-     * Create a new CommitteeMember. Only available as admin.
-     *
-     * @param committeeMember The CommitteeMember to add.
-     * @return The status of the addition.
-     */
-    @PreAuthorize("hasRole('ADMIN')")
-    @PostMapping("/committee")
-    public ResponseEntity<?> addCommitteeMember(@RequestBody CommitteeMember committeeMember) {
-        webService.addCommitteeMember(committeeMember);
-        return ResponseEntityBuilder.createResponseEntity(HttpStatus.CREATED, "Committee member added successfully.");
-    }
-
-    /**
-     * Retrieve all CommitteeMembers.
-     *
-     * @return All CommitteeMembers.
-     */
-    @GetMapping("/committee")
-    public ResponseEntity<?> readCommittee() {
-        Collection<CommitteeMember> committeeMembers = webService.getCommittee();
-        return ResponseEntityBuilder.createResponseEntity(HttpStatus.OK, "Committee members retrieved successfully.", committeeMembers);
-    }
-
-    /**
-     * Update the CommitteeMember at the given id with the given values.
-     * Only available as admin.
-     *
-     * @param id The id at which to update the CommitteeMember.
-     * @param committeeMember The new Values for the CommitteeMember.
-     * @return The status of the update.
-     */
-    @PreAuthorize("hasRole('ADMIN')")
-    @PutMapping("/committee/{memberID}")
-    public ResponseEntity<?> addCommitteeMember(@PathVariable Long id, @RequestBody CommitteeMember committeeMember) {
-        webService.updateCommitteeMember(id, committeeMember);
-        return ResponseEntityBuilder.createResponseEntity(HttpStatus.ACCEPTED, "Committee member updated successfully.");
-    }
-
-    /**
-     * Delete a CommitteeMember. Only available as admin.
-     *
-     * @param id the id of the CommitteeMember to be deleted.
-     * @return The status of the deletion.
-     */
-    @PreAuthorize("hasRole('ADMIN')")
-    @DeleteMapping("/committee/{memberID}")
-    public ResponseEntity<?> deleteCommitteeMember(@PathVariable Long id) {
-        webService.deleteCommitteeMember(id);
-        return ResponseEntityBuilder.createResponseEntity(HttpStatus.OK, "Committee member deleted successfully.");
-    }
 }
